Adaptation to Ginkgo v1
